// src/views/backend-manage/table/read-xls/config/type.ts
export interface DataType {
  key: string;
  name: string;
  age: number;
  height: number;
  weight: number;
  address: string;
  political: string;
  education: string;
  school: string;
}

export interface IXLSXData {
  住址: string;
  "体重(kg)": number;
  姓名: string;
  学历: string;
  学校: string;
  年龄: number;
  政治面貌: string;
  "身高(m)": number;
}